Tested MD app, set up local-runs infrastructure
When testing each app locally, it is useful to sandbox the
app to see exactly which files are inputs and outputs, without
the other apps running.  Here, I setup a new directory for each
app step in the workflow (e.g. step_01) and tested steps 1 and 2
in this framework. I also had to add execute permissions to
scripts/binaries in `../models` and its subdirectories.