Trying to speed up workflow
Not using convert means that larger (1.4MB) image
files are sent back to the platform.  This bigger data
transfer slows things down.  Also, I think we need to
verify which files/paths are inputs and which are outputs
to the visualization app.  If a path is set as an input, it
looks like execution is serial if there are multiple apps
all touching the same path.  Could try reversing the loop order
(so all cases only have one worker on them, rather than several
workers waiting to work on the same case).