Test staging specific files
The visualization stage of the workflow runs very slowly b/c
the Path() calls may require rsyncing and workflow pausing for
every time there is a change to files in that directory.  The
goal is for each frame in an animation to be created separately
from the other frames.  However, workers were pulling all
previously rendered frames from PW, which caused a big IO
bottleneck. This version uses Path() to stage single files
and it behaves as expected (only staging needed files) but
it's still pretty slow and there is no speedup when the
workflow uses 2x more workers.