First draft of MDLite with std Parsl
Initial testing is with AECM cluster. Currently not working due
to workers not able to find a viable address to connect back to
the interchange.  Need to find the IP or hostname of the user_host
visible to the worker nodes. The interchange is running on the
user_host, which logs into the submit_host, workers need to connect
back to the Parsl Interchange on the user_host. Parsl CAN stage
SLURM submit scripts and process_worker_pool.py is found and started.