Modified vis tool for shared file system
The visualization tool in this workflow, C-ray, depends on a
configuration file, `c-ray-trailer` and is launched from a runscript
`renderframe`.  So that C-ray can run on a shared filesystem with
standard Parsl's NoOpFileStaging, I could not depend on each C-ray
invocation to be in it's own siloed task directory. This meant:
1. creating `c-ray-trailer` as a persistent, shared file
2. ...with `make_c-ray-trailer.sh` for documentation, and
3. making a new `renderframe_shared_fs`.
In particular the older `renderframe` used relative paths and for the
shared filesystem, `renderframe_shared_fs` uses absolute paths.
The shared file system version is run with `main_std_parsl.ipynb`
while the PW-Parsl version (with siloed task dirs) is in `main.ipynb`.