Require 1 core per worker -> impact on rsync!
The rsync limits apply to the number of multiplex connections that
can be used on both the head node and worker nodes, depending on
where the rsync is called.  This means that workers as well as the
head nodes need MaxSessions boosted in /etc/ssh/sshd_config.
However, since the workers are ephemeral, need to apply an automated
config change to ensure this setting is present.  Here, the work
around is to require 1 CPU per worker (up from 0.1 CPU per worker)
which reduces the number of tasks that are able to run on a particular
worker node. The fewer the number of running tasks on workers, the
less likely we are going to come up against the MaxSessions limit.
Fixing MaxSessions is better, but try this for now to verify the
expected number of workers, etc.