Max blocks >= number tasks will one-to-one tasks/jobs
...but if the number of available worker NODES is less than the number
of tasks, then multiple tasks are sent to Parsl workers (i.e. Parsl
workers are pilot jobs that will keep accepting tasks even though
there are other workers trying to start in the queue -> this means
that not all Parsl workers are necessarily going to get tasks assigned
to them). We therefore need a way to:
1) enforce one and only one task per worker, and
2) find out what IDs the SlurmProvider.cancel() method can use to stop jobs on-the-fly.